refactor to split up groupby
