+++ modules/cudf/src/groupby.ts
@@ -12,467 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import {Series, Struct} from '@nvidia/cudf';
-import {MemoryResource} from '@nvidia/rmm';
-import {Field} from 'apache-arrow';
-
-import CUDF from './addon';
-import {Column} from './column';
-import {DataFrame, SeriesMap} from './data_frame';
-import {Table} from './table';
-import {DataType} from './types/dtypes';
-import {NullOrder} from './types/enums'
-import {Interpolation, TypeMap} from './types/mappings'
-
-/*
- * @param keys DataFrame whose rows act as the groupby keys
- * @param include__nulls Indicates whether rows in `keys` that contain
- * NULL values should be included
- * @param keys_are_sorted Indicates whether rows in `keys` are already sorted
- * @param column_order If `keys_are_sorted == YES`, indicates whether each
- * column is ascending/descending. If empty, assumes all  columns are
- * ascending. Ignored if `keys_are_sorted == false`.
- * @param null_precedence If `keys_are_sorted == YES`, indicates the ordering
- * of null values in each column. Else, ignored. If empty, assumes all columns
- * use `null_order::BEFORE`. Ignored if `keys_are_sorted == false`.
- */
-
-type TypeMapOf<T extends string, R extends DataType> = {
-  [P in T]: R
-};
-
-type GroupByBaseProps = {
-  include_nulls?: boolean,
-  keys_are_sorted?: boolean,
-  column_order?: boolean[],
-  null_precedence?: NullOrder[],
-}
-
-export type GroupBySingleProps<T extends TypeMap, R extends keyof T> = {
-  obj: DataFrame<T>,
-  by: R,
-}&GroupByBaseProps;
-
-export type GroupByMultipleProps<T extends TypeMap, R extends keyof T, IndexKey extends string> = {
-  obj: DataFrame<T>,
-  by: R[],
-  index_key: IndexKey,
-}&GroupByBaseProps;
-
-type CudfGroupByProps = {
-  keys: Table,
-}&GroupByBaseProps;
-
-export type Groups<KeysMap extends TypeMap, ValuesMap extends TypeMap> = {
-  keys: DataFrame<KeysMap>,
-  offsets: Int32Array,
-  values?: DataFrame<ValuesMap>,
-}
-
-interface CudfGroupBy {
-  _getGroups(values?: Table,
-             memoryResource?: MemoryResource): {keys: Table, offsets: Int32Array, values?: Table};
-
-  _argmax(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _argmin(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _count(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _max(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _mean(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _median(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _min(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _nth(n: number, values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _nunique(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _std(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _sum(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _var(values: Table, memoryResource?: MemoryResource): {keys: Table, cols: Column[]};
-  _quantile(q: number, values: Table, interpolation?: number, memoryResource?: MemoryResource):
-    {keys: Table, cols: [Column]};
-}
-
-class GroupByBase<T extends TypeMap, R extends keyof T> {
-  protected _by: R[];
-  protected _values: DataFrame<Omit<T, R>>;
-  protected _cudf_groupby: CudfGroupBy;
-
-  constructor(props: GroupByBaseProps, by: R[], obj: DataFrame<T>) {
-    const table = obj.select(by).asTable();
-
-    const {
-      include_nulls   = false,
-      keys_are_sorted = false,
-      column_order    = [],
-      null_precedence = []
-    } = props;
-
-    const cudf_props: CudfGroupByProps = {
-      keys: table,
-      include_nulls: include_nulls,
-      keys_are_sorted: keys_are_sorted,
-      column_order: column_order,
-      null_precedence: null_precedence,
-    };
-
-    this._cudf_groupby = new CUDF.GroupBy(cudf_props);
-
-    this._by     = by;
-    this._values = obj.drop(by);
-  }
-
-  /**
-   * Return the Groups for this GroupBy
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  getGroups(memoryResource?: MemoryResource) {
-    const {keys, offsets, values} =
-      this._cudf_groupby._getGroups(this._values.asTable(), memoryResource);
-
-    const results = {
-      offsets,
-      keys: new DataFrame(
-        this._by.reduce((keys_map, name, index) =>
-                          ({...keys_map, [name]: Series.new(keys.getColumnByIndex(index))}),
-                        {} as SeriesMap<Pick<T, R>>))
-    } as Groups<Pick<T, R>, Omit<T, R>>;
-
-    if (values !== undefined) {
-      results.values = new DataFrame(this._values.names.reduce(
-        (values_map, name, index) =>
-          ({...values_map, [name]: Series.new(values.getColumnByIndex(index))}),
-        {} as SeriesMap<Omit<T, R>>));
-    }
-
-    return results;
-  }
-}
-
-export class GroupBySingle<T extends TypeMap, R extends keyof T> extends GroupByBase<T, R> {
-  constructor(props: GroupBySingleProps<T, R>) { super(props, [props.by], props.obj); }
-
-  protected prepare_results(results: {keys: Table, cols: Column[]}) {
-    const {keys, cols} = results;
-    const series_map   = {} as SeriesMap<T>;
-
-    this._values.names.forEach((name, index) => { series_map[name] = Series.new(cols[index]); });
-    series_map[this._by[0]] = Series.new(keys.getColumnByIndex(0));
-
-    return new DataFrame(series_map);
-  }
-
-  /**
-   * Compute the index of the maximum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  argmax(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._argmax(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the index of the minimum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  argmin(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._argmin(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the size of each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  count(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._count(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the maximum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  max(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._max(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the average value each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  mean(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._mean(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the median value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  median(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._median(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the minimum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  min(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._min(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Return the nth value from each group
-   *
-   * @param n the index of the element to return
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  nth(n: number, memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._nth(n, this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the number of unique values in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  nunique(memoryResource?: MemoryResource) {
-    return this.prepare_results(
-      this._cudf_groupby._nunique(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the standard deviation for each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  std(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._std(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the sum of values in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  sum(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._sum(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the variance for each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  var(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._var(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Return values at the given quantile.
-   *
-   * @param q the quantile to compute, 0 <= q <= 1
-   * @param interpolation This optional parameter specifies the interpolation method to use,
-   *  when the desired quantile lies between two data points i and j.
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  quantile(q                                         = 0.5,
-           interpolation: keyof typeof Interpolation = 'linear',
-           memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._quantile(
-      q, this._values.asTable(), Interpolation[interpolation], memoryResource));
-  }
-}
-
-export class GroupByMultiple<T extends TypeMap, R extends keyof T, IndexKey extends string> extends
-  GroupByBase<T, R> {
-  private index_key: IndexKey;
-  constructor(props: GroupByMultipleProps<T, R, IndexKey>) {
-    super(props, props.by, props.obj);
-    this.index_key = props.index_key;
-  }
-
-  protected prepare_results(results: {keys: Table, cols: Column[]}) {
-    const {keys, cols} = results;
-
-    type Subset        = Pick<T, R>;
-    type Index         = Struct<Subset>;
-    type IndexTypeMap  = TypeMapOf<IndexKey, Index>;
-    type RestTypeMap   = Omit<T, R>;
-    type RestSeriesMap = SeriesMap<RestTypeMap>;
-
-    const rest_map = this._values.names.reduce(
-      (xs, key, index) => ({...xs, [key]: Series.new(cols[index])}), {} as RestSeriesMap);
-
-    if (this.index_key in rest_map) {
-      throw new Error(`Groupby column name ${this.index_key} already exists`);
-    }
-
-    const fields   = [];
-    const children = [];
-    for (const [index, name] of this._by.entries()) {
-      const child = keys.getColumnByIndex(index)
-      fields.push(Field.new({name: name as string, type: child.type}));
-      const series = Series.new(child);
-      children.push(series);
-    }
-
-    const index = Series.new<Index>({type: new Struct(fields), children: children});
-
-    return new DataFrame({[this.index_key]: index, ...rest_map} as
-                         SeriesMap<IndexTypeMap&RestTypeMap>);
-  }
-
-  /**
-   * Compute the index of the maximum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  argmax(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._argmax(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the index of the minimum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  argmin(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._argmin(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the size of each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  count(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._count(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the maximum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  max(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._max(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the average value each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  mean(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._mean(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the median value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  median(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._median(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the minimum value in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  min(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._min(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Return the nth value from each group
-   *
-   * @param n the index of the element to return
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  nth(n: number, memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._nth(n, this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the number of unique values in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  nunique(memoryResource?: MemoryResource) {
-    return this.prepare_results(
-      this._cudf_groupby._nunique(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the standard deviation for each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  std(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._std(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the sum of values in each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  sum(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._sum(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Compute the variance for each group
-   *
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  var(memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._var(this._values.asTable(), memoryResource));
-  }
-
-  /**
-   * Return values at the given quantile.
-   *
-   * @param q the quantile to compute, 0 <= q <= 1
-   * @param interpolation This optional parameter specifies the interpolation method to use,
-   *  when the desired quantile lies between two data points i and j.
-   * @param memoryResource The optional MemoryResource used to allocate the result's
-   *   device memory.
-   */
-  quantile(q                                         = 0.5,
-           interpolation: keyof typeof Interpolation = 'linear',
-           memoryResource?: MemoryResource) {
-    return this.prepare_results(this._cudf_groupby._quantile(
-      q, this._values.asTable(), Interpolation[interpolation], memoryResource));
-  }
-}
+export {GroupByBaseProps} from './groupby/base';
+export {GroupByMultiple, GroupByMultipleProps} from './groupby/multiple'
+export {GroupBySingle, GroupBySingleProps} from './groupby/single'
